+++ src/ParticleState.cpp
@@ -49,14 +49,12 @@ double ParticleState::getRigidity() const {
 
 void ParticleState::setId(int newId) {
 	id = newId;
+	pmass = particleMass(id);
 	if (isNucleus(id)) {
-		pmass = nuclearMass(id);
 		charge = chargeNumber(id) * eplus;
 		if (id < 0)
 			charge *= -1; // anti-nucleus
 	} else {
-		if (abs(id) == 11)
-			pmass = mass_electron;
 		charge = HepPID::charge(id) * eplus;
 	}
 }